+++ integration/batching-with-context/batching.ts
@@ -116,7 +116,7 @@ export class EntityServiceClientImpl<Context extends DataLoaders> implements Ent
       return new DataLoader<string, Entity>((ids) => {
         const request = { ids };
         return this.BatchQuery(ctx, request).then(res => res.entities);
-      }, { cacheKeyFn: hash });
+      }, { cacheKeyFn: hash, ...ctx.rpcDataLoaderOptions });
     });
     return dl.load(id);
   }
@@ -134,7 +134,7 @@ export class EntityServiceClientImpl<Context extends DataLoaders> implements Ent
         return this.BatchMapQuery(ctx, request).then(res => {
           return ids.map(key => res.entities[key]);
         })
-      }, { cacheKeyFn: hash });
+      }, { cacheKeyFn: hash, ...ctx.rpcDataLoaderOptions });
     });
     return dl.load(id);
   }
@@ -154,7 +154,7 @@ export class EntityServiceClientImpl<Context extends DataLoaders> implements Ent
           return GetOnlyMethodResponse.decode(new Reader(response));
         })
         return Promise.all(responses);
-      }, { cacheKeyFn: hash });
+      }, { cacheKeyFn: hash, ...ctx.rpcDataLoaderOptions  });
     });
     return dl.load(request);
   }
@@ -173,11 +173,18 @@ interface Rpc<Context> {
 
 }
 
+export interface DataLoaderOptions {
+  cache?: boolean;
+}
+
 export interface DataLoaders {
+  rpcDataLoaderOptions?: DataLoaderOptions;
 
   getDataLoader<T>(identifier: string, constructorFn: () => T): T;
 
 }
+
+export const protobufPackage = 'batching'
 
 export const BatchQueryRequest = {
   encode(message: BatchQueryRequest, writer: Writer = Writer.create()): Writer {
@@ -711,7 +718,7 @@ export const Entity = {
 };
 
 type Builtin = Date | Function | Uint8Array | string | number | undefined;
-type DeepPartial<T> = T extends Builtin
+export type DeepPartial<T> = T extends Builtin
   ? T
   : T extends Array<infer U>
   ? Array<DeepPartial<U>>