+++ integration/only-types/google/protobuf/timestamp.ts
@@ -100,3 +100,5 @@ export interface Timestamp {
    */
   nanos: number;
 }
+
+export const protobufPackage = 'google.protobuf'