+++ integration/simple-long/simple.ts
@@ -5,7 +5,7 @@ import { ImportedThing } from './import_dir/thing';
 import * as Long from 'long';
 import { Reader, Writer, util, configure } from 'protobufjs/minimal';
 import { Timestamp } from './google/protobuf/timestamp';
-import { StringValue, Int32Value, BoolValue } from './google/protobuf/wrappers';
+import { StringValue, Int32Value, BoolValue, Int64Value } from './google/protobuf/wrappers';
 
 
 /**
@@ -66,6 +66,7 @@ export interface SimpleWithWrappers {
   name: string | undefined;
   age: number | undefined;
   enabled: boolean | undefined;
+  bananas: Long | undefined;
   coins: number[];
   snacks: string[];
 }
@@ -264,6 +265,8 @@ function fromTimestamp(t: Timestamp): Date {
 function numberToLong(number: number) {
   return Long.fromNumber(number);
 }
+
+export const protobufPackage = 'simple'
 
 export enum StateEnum {
   UNKNOWN = 0,
@@ -950,6 +953,9 @@ export const SimpleWithWrappers = {
     if (message.enabled !== undefined && message.enabled !== undefined) {
       BoolValue.encode({ value: message.enabled! }, writer.uint32(26).fork()).ldelim();
     }
+    if (message.bananas !== undefined && message.bananas !== undefined) {
+      Int64Value.encode({ value: message.bananas! }, writer.uint32(34).fork()).ldelim();
+    }
     for (const v of message.coins) {
       Int32Value.encode({ value: v!! }, writer.uint32(50).fork()).ldelim();
     }
@@ -975,6 +981,9 @@ export const SimpleWithWrappers = {
           break;
         case 3:
           message.enabled = BoolValue.decode(reader, reader.uint32()).value;
+          break;
+        case 4:
+          message.bananas = Int64Value.decode(reader, reader.uint32()).value;
           break;
         case 6:
           message.coins.push(Int32Value.decode(reader, reader.uint32()).value);
@@ -1008,6 +1017,11 @@ export const SimpleWithWrappers = {
     } else {
       message.enabled = undefined;
     }
+    if (object.bananas !== undefined && object.bananas !== null) {
+      message.bananas = Long.fromValue(object.bananas);
+    } else {
+      message.bananas = undefined;
+    }
     if (object.coins !== undefined && object.coins !== null) {
       for (const e of object.coins) {
         message.coins.push(Number(e));
@@ -1039,6 +1053,11 @@ export const SimpleWithWrappers = {
     } else {
       message.enabled = undefined;
     }
+    if (object.bananas !== undefined && object.bananas !== null) {
+      message.bananas = object.bananas as Long | undefined;
+    } else {
+      message.bananas = undefined;
+    }
     if (object.coins !== undefined && object.coins !== null) {
       for (const e of object.coins) {
         message.coins.push(e);
@@ -1056,6 +1075,7 @@ export const SimpleWithWrappers = {
     message.name !== undefined && (obj.name = message.name);
     message.age !== undefined && (obj.age = message.age);
     message.enabled !== undefined && (obj.enabled = message.enabled);
+    message.bananas !== undefined && (obj.bananas = message.bananas);
     if (message.coins) {
       obj.coins = message.coins.map(e => e);
     } else {
@@ -1865,7 +1885,7 @@ if (util.Long !== Long as any) {
 }
 
 type Builtin = Date | Function | Uint8Array | string | number | undefined;
-type DeepPartial<T> = T extends Builtin
+export type DeepPartial<T> = T extends Builtin
   ? T
   : T extends Array<infer U>
   ? Array<DeepPartial<U>>