+++ integration/simple-proto2/simple.ts
@@ -9,6 +9,8 @@ const baseIssue56: object = {
   test: 1,
 };
 
+export const protobufPackage = 'simple'
+
 export enum EnumWithoutZero {
   A = 1,
   B = 2,
@@ -89,7 +91,7 @@ export const Issue56 = {
 };
 
 type Builtin = Date | Function | Uint8Array | string | number | undefined;
-type DeepPartial<T> = T extends Builtin
+export type DeepPartial<T> = T extends Builtin
   ? T
   : T extends Array<infer U>
   ? Array<DeepPartial<U>>