+++ integration/simple-snake/simple.ts
@@ -268,6 +268,8 @@ function longToNumber(long: Long) {
   return long.toNumber();
 }
 
+export const protobufPackage = 'simple'
+
 export enum StateEnum {
   UNKNOWN = 0,
   ON = 2,
@@ -1868,7 +1870,7 @@ if (util.Long !== Long as any) {
 }
 
 type Builtin = Date | Function | Uint8Array | string | number | undefined;
-type DeepPartial<T> = T extends Builtin
+export type DeepPartial<T> = T extends Builtin
   ? T
   : T extends Array<infer U>
   ? Array<DeepPartial<U>>