+++ integration/simple-string-enums/simple.ts
@@ -1,0 +1,15 @@
+
+
+export interface Simple {
+  name: string;
+  state: StateEnum;
+}
+
+export const protobufPackage = 'simple'
+
+export enum StateEnum {
+  UNKNOWN = "UNKNOWN",
+  ON = "ON",
+  OFF = "OFF",
+  UNRECOGNIZED = "UNRECOGNIZED",
+}