+++ integration/vector-tile/vector_tile.ts
@@ -66,6 +66,8 @@ function longToNumber(long: Long) {
   return long.toNumber();
 }
 
+export const protobufPackage = 'vector_tile'
+
 export enum Tile_GeomType {
   UNKNOWN = 0,
   POINT = 1,
@@ -586,7 +588,7 @@ if (util.Long !== Long as any) {
 }
 
 type Builtin = Date | Function | Uint8Array | string | number | undefined;
-type DeepPartial<T> = T extends Builtin
+export type DeepPartial<T> = T extends Builtin
   ? T
   : T extends Array<infer U>
   ? Array<DeepPartial<U>>